+++ projects/angular-toolbox-demo/public/documentation/laf-less-components/index.html
@@ -163,13 +163,25 @@
             </tr>
             <tr>
                 <td>
-                    <strong>Popup Container</strong>
+                    <strong>Dialog API</strong>
                 </td>
                 <td>
-                    Not available yet.
+                    Creates modal popups that display custom Angular components.
                 </td>
-                <td></td>
-                <td></td>
+                <td>
+                    <button class="btn btn-outline-primary" onclick="appBridge.navigate(['resources/documentation/dialog-api'])" title="Dialog API">
+                        <svg xmlns="http://www.w3.org/2000/svg" width="20" height="20">
+                            <use href="#linkIcon"/>
+                        </svg>
+                    </button>
+                </td>
+                <td>
+                    <button class="btn btn-outline-secondary" onclick="appBridge.navigate(['demo/dialog'])" title="Dialog Service Demo">
+                        <svg xmlns="http://www.w3.org/2000/svg" width="20" height="20">
+                            <path d="M10.478 1.647a.5.5 0 1 0-.956-.294l-4 13a.5.5 0 0 0 .956.294zM4.854 4.146a.5.5 0 0 1 0 .708L1.707 8l3.147 3.146a.5.5 0 0 1-.708.708l-3.5-3.5a.5.5 0 0 1 0-.708l3.5-3.5a.5.5 0 0 1 .708 0m6.292 0a.5.5 0 0 0 0 .708L14.293 8l-3.147 3.146a.5.5 0 0 0 .708.708l3.5-3.5a.5.5 0 0 0 0-.708l-3.5-3.5a.5.5 0 0 0-.708 0"/>
+                        </svg>
+                    </button>
+                </td>
             </tr>
         </tbody>
       </table>
