+++ projects/angular-toolbox/src/lib/component/index.ts
@@ -9,4 +9,3 @@
 export * from './angular-toolbox-logo/angular-toolbox-logo.component';
 export * from './dropdown/dropdown.component';
 export * from './layout';
-export * from './dialog';
