+++ projects/angular-toolbox/src/lib/framework/dialog/dialog-backdrop.type.ts
