+++ projects/angular-toolbox/src/lib/framework/dialog/dialog-default-config.ts
