+++ projects/angular-toolbox/src/lib/framework/dialog/dialog-service.ts
