+++ projects/angular-toolbox/src/lib/framework/dialog/dialog.config.ts
