+++ projects/angular-toolbox/src/lib/framework/dialog/dialog.outlet.ts
