+++ projects/angular-toolbox/src/lib/framework/index.ts
@@ -3,3 +3,4 @@ export * from './mock/http';
 export * from './mock/http-monitoring-console';
 export * from './mock/documentation';
 export * from './logging';
+export * from './dialog';
