+++ projects/angular-toolbox/src/test/framework/dialog/dialog-default-config.spec.ts
@@ -6,7 +6,7 @@
  * found in the LICENSE file at https://pascalechemann.com/angular-toolbox/resources/license
  */
 
-import { DEFAULT_CONFIG } from "projects/angular-toolbox/src/lib/component/dialog/dialog-default-config";
+import { DEFAULT_CONFIG } from "projects/angular-toolbox/src/lib/framework/dialog/dialog-default-config";
 import { DialogBackdropType } from "projects/angular-toolbox/src/public-api";
 
 describe('DEFAULT_CONFIG', () => {
